Changing : saving all checkin and filtering in get

+++ src/main/java/fr/socrates/domain/checkin/Printer.java
@@ -1,5 +1,5 @@
 package fr.socrates.domain.checkin;
 
 interface Printer {
-    String print(int coldFoodCount);
+    String print(long coldFoodCount);
 }
